compile fix

+++ src/OpenCOVER/cover/coVRConfig.cpp
@@ -293,7 +293,7 @@ coVRConfig::coVRConfig()
         float h, p, r;
         
         char str[200];
-        sprintf(str, "COVER.ScreenConfig.Screen:%d", i);
+        sprintf(str, "COVER.ScreenConfig.Screen:%d", (int)i);
         bool state = coCoverConfig::getScreenConfigEntry(i, screens[i].name, &hsize, &vsize, &x, &y, &z, &h, &p, &r);
         if (!state)
         {
@@ -328,7 +328,7 @@ coVRConfig::coVRConfig()
     for (size_t i = 0; i < pipes.size(); i++)
     {
         char str[200];
-        sprintf(str, "COVER.PipeConfig.Pipe:%d", i);
+        sprintf(str, "COVER.PipeConfig.Pipe:%d", (int)i);
         pipes[i].x11DisplayNum = coCoviseConfig::getInt("server", str, 0);
         pipes[i].x11ScreenNum = coCoviseConfig::getInt("screen", str, 0);
     }
@@ -351,7 +351,7 @@ coVRConfig::coVRConfig()
         std::string stereoM;
 
         char str[200];
-        sprintf(str, "COVER.ChannelConfig.Channel:%d", i);
+        sprintf(str, "COVER.ChannelConfig.Channel:%d", (int)i);
         std::string s = coCoviseConfig::getEntry("comment", str, "NoNameChannel");
         channels[i].name = s;
         stereoM = coCoviseConfig::getEntry("stereoMode", str);
@@ -400,7 +400,7 @@ coVRConfig::coVRConfig()
         std::string stereoM;
 
         char str[200];
-        sprintf(str, "COVER.PBOConfig.PBO:%d", i);
+        sprintf(str, "COVER.PBOConfig.PBO:%d", (int)i);
         
         PBOs[i].PBOsx = coCoviseConfig::getInt("PBOSizeX", str, -1);
         PBOs[i].PBOsy = coCoviseConfig::getInt("PBOSizeY", str, -1);
@@ -411,7 +411,7 @@ coVRConfig::coVRConfig()
         std::string stereoM;
 
         char str[200];
-        sprintf(str, "COVER.ViewportConfig.Viewport:%d", i);
+        sprintf(str, "COVER.ViewportConfig.Viewport:%d", (int)i);
         viewportStruct &vp = viewports[i];
         bool exists=false;
         vp.window = coCoviseConfig::getInt("windowIndex", str, -1,&exists);
@@ -419,7 +419,7 @@ coVRConfig::coVRConfig()
         {
             // no viewport config, check for values in channelConfig for backward compatibility
             
-            sprintf(str, "COVER.ChannelConfig.Channel:%d", i);
+            sprintf(str, "COVER.ChannelConfig.Channel:%d", (int)i);
             vp.window = coCoviseConfig::getInt("windowIndex", str, -1,&exists);
             if (!exists)
             {
@@ -532,7 +532,7 @@ coVRConfig::coVRConfig()
     for (size_t i = 0; i < blendingTextures.size(); i++)
     {
         char str[200];
-        sprintf(str, "COVER.BlendingTextureConfig.BlendingTexture:%d", i);
+        sprintf(str, "COVER.BlendingTextureConfig.BlendingTexture:%d", (int)i);
         blendingTextureStruct &bt = blendingTextures[i];
         bool exists=false;
         bt.window = coCoviseConfig::getInt("windowIndex", str, -1,&exists);
